+++ health/api/src/main/java/com/cerner/beadledom/health/dto/HealthDto.java
@@ -28,6 +28,7 @@ import javax.annotation.Nullable;
 @AutoValue
 @JsonDeserialize(builder = HealthDto.Builder.class)
 @ApiModel(description = "Indicates the health of this service.")
+@io.swagger.annotations.ApiModel(description = "Indicates the health of this service.")
 public abstract class HealthDto {
   /**
    * Creates a new builder for {@code HealthDto}.
@@ -76,6 +77,8 @@ public abstract class HealthDto {
    * information.
    */
   @ApiModelProperty(value = "a human-readable status description", required = true)
+  @io.swagger.annotations.ApiModelProperty(value = "a human-readable status description",
+      required = true)
   @JsonProperty("message")
   @JsonView(Availability.class)
   public abstract Optional<String> getMessage();
@@ -91,17 +94,22 @@ public abstract class HealthDto {
 
   @ApiModelProperty(
       "the results of any dependency health checks invoked as part of this health check")
+  @io.swagger.annotations.ApiModelProperty(
+      "the results of any dependency health checks invoked as part of this health check")
   @JsonProperty("dependencies")
   @JsonView({Primary.class, Diagnostic.class, HealthJsonViews.Dependency.class})
   public abstract Optional<List<HealthDependencyDto>> getDependencies();
 
   @ApiModelProperty(
       "Build and version information of the service serving the health check response")
+  @io.swagger.annotations.ApiModelProperty(
+      "Build and version information of the service serving the health check response")
   @JsonProperty("build")
   @JsonView(Diagnostic.class)
   public abstract Optional<BuildDto> getBuild();
 
   @ApiModelProperty("Runtime and environment information of the server")
+  @io.swagger.annotations.ApiModelProperty("Runtime and environment information of the server")
   @JsonProperty("server")
   @JsonView(Diagnostic.class)
   public abstract Optional<ServerDto> getServer();