+++ pagination/src/main/java/com/cerner/beadledom/pagination/parameters/LimitParameter.java
@@ -2,9 +2,8 @@ package com.cerner.beadledom.pagination.parameters;
 
 import com.cerner.beadledom.pagination.OffsetPaginationModule;
 import com.cerner.beadledom.pagination.models.OffsetPaginationConfiguration;
-import com.wordnik.swagger.annotations.ApiModel;
-import com.wordnik.swagger.annotations.ApiModelProperty;
-import javax.inject.Inject;
+import com.google.inject.Inject;
+import com.wordnik.swagger.annotations.ApiParam;
 
 /**
  * Represent the limit parameter used for pagination.
@@ -18,10 +17,10 @@ import javax.inject.Inject;
  * @author Ian Kottman
  * @since 3.1
  */
-@ApiModel
 public class LimitParameter extends AbstractParameter<Integer> {
 
-  @ApiModelProperty(value = "Total number of items to return in the response.", dataType = "int")
+  @ApiParam(value = "Total number of items to return in the response.")
+  @io.swagger.annotations.ApiParam(value = "Total number of items to return in the response.")
   private final String limit;
 
   @Inject